Move api calls into their own modules
